v2.0.30 | 2019-12-20 | Add option to keep database option on uninstall in case of upgrading to PRO version

+++ wp-foft-loader.php
@@ -2,7 +2,7 @@
 
 /**
  * Plugin Name: WP FOFT Loader
- * Version: 2.0.29
+ * Version: 2.0.30
  * Author URI: https://github.com/seezee
  * Plugin URI: https://wordpress.org/plugins/wp-foft-loader/
  * GitHub Plugin URI: seezee/WP-FOFT-Loader  
@@ -77,7 +77,7 @@ if ( !defined( '_WPFL_BASE_' ) ) {
 
 
 if ( !defined( '_WPFL_VERSION_' ) ) {
-    define( '_WPFL_VERSION_', '2.0.29' );
+    define( '_WPFL_VERSION_', '2.0.30' );
 } else {
     echo  '<div id="updated" class="notice notice-error is-dismissible"><span class="dashicons dashicons-no"></span> ' . __( 'WP <abb>FOFT</abbr> Loader ERROR! The <abbr>PHP</abbr> constant', 'wp-foft-loader' ) . ' &ldquo;_WPFL_VERSION_&rdquo; ' . __( 'has already been defined. This could be due to a conflict with another plugin or theme. Please check your logs to debug.', 'wp-foft-loader' ) . '</div>' ;
 }
@@ -165,11 +165,24 @@ function wpfl_check_version()
 add_action( 'plugins_loaded', 'wpfl_check_version' );
 function wpfl_fs_uninstall_cleanup()
 {
-    foreach ( wp_load_alloptions() as $option => $value ) {
-        if ( strpos( $option, _WPFL_BASE_ ) === 0 ) {
-            delete_option( $option );
+    
+    if ( get_option( _WPFL_BASE_ . 'uninstall' ) !== NULL ) {
+        $uninstall = get_option( _WPFL_BASE_ . 'uninstall' );
+    } else {
+        $uninstall = NULL;
+    }
+    
+    
+    if ( $uninstall !== NULL && $uninstall !== 'delete' ) {
+        return;
+    } else {
+        foreach ( wp_load_alloptions() as $option => $value ) {
+            if ( strpos( $option, _WPFL_BASE_ ) === 0 ) {
+                delete_option( $option );
+            }
         }
     }
+
 }
 
 // Uninstall hook.
